database/database.py
# database/database.py
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker, Session
from dotenv import load_dotenv
import os

# .env 로드
load_dotenv()

DB_URL = os.getenv("DB_URL")
if not DB_URL:
    raise ValueError("❌ DB_URL not found in environment variables")

# 엔진 생성
engine = create_engine(
    DB_URL,
    pool_size=10,
    max_overflow=20,
    pool_pre_ping=True,
)

# 세션 팩토리
SessionLocal = sessionmaker(
    autocommit=False,
    autoflush=False,
    bind=engine
)

# ORM 모델 Base
Base = declarative_base()


# ✅ 여기 추가
def get_db():
    db: Session = SessionLocal()
    try:
        yield db
    finally:
        db.close()
